Extract selection logit as hook
